feat：支持跨项目复制流水线 #12918 删除多余的代码
