+++ src/backend/ci/core/process/api-process/src/main/kotlin/com/tencent/devops/process/api/op/OpPipelineViewResource.kt
@@ -10,7 +10,6 @@ import jakarta.ws.rs.Consumes
 import jakarta.ws.rs.DELETE
 import jakarta.ws.rs.GET
 import jakarta.ws.rs.HeaderParam
-import jakarta.ws.rs.POST
 import jakarta.ws.rs.Path
 import jakarta.ws.rs.PathParam
 import jakarta.ws.rs.Produces
@@ -44,23 +43,5 @@ interface OpPipelineViewResource {
         @Parameter(description = "yaml文件目录", required = true)
         @QueryParam("directory")
         directory: String
-    ): Result<Boolean>
-
-    @Operation(summary = "跨项目复制流水线组")
-    @POST
-    @Path("/projects/{sourceProjectId}/copyAcrossProject")
-    fun copyAcrossProject(
-        @Parameter(description = "用户ID", required = true, example = AUTH_HEADER_USER_ID_DEFAULT_VALUE)
-        @HeaderParam(AUTH_HEADER_USER_ID)
-        userId: String,
-        @Parameter(description = "源项目ID", required = true)
-        @PathParam("sourceProjectId")
-        sourceProjectId: String,
-        @Parameter(description = "目标项目ID", required = true)
-        @QueryParam("targetProjectId")
-        targetProjectId: String,
-        @Parameter(description = "源流水线组名称，为空则复制源项目下全部流水线组", required = false)
-        @QueryParam("viewName")
-        viewName: String?
     ): Result<Boolean>
 }
