+++ src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/service/pipeline/PipelineCopyService.kt
@@ -1,19 +1,26 @@
 package com.tencent.devops.process.service.pipeline
 
 import com.tencent.devops.auth.api.service.ServiceResourceMemberResource
+import com.tencent.devops.common.api.util.HashUtil
 import com.tencent.devops.common.auth.api.AuthResourceType
 import com.tencent.devops.common.client.Client
 import com.tencent.devops.common.client.ClientTokenService
-import com.tencent.devops.common.pipeline.enums.ChannelCode
-import com.tencent.devops.common.pipeline.enums.PipelineInstanceTypeEnum
-import com.tencent.devops.common.pipeline.enums.VersionStatus
 import com.tencent.devops.common.pipeline.pojo.setting.PipelineSetting
-import com.tencent.devops.model.process.tables.records.TPipelineInfoRecord
+import com.tencent.devops.model.process.tables.records.TPipelineLabelRecord
+import com.tencent.devops.model.process.tables.records.TPipelineViewRecord
+import com.tencent.devops.process.constant.PipelineViewType
 import com.tencent.devops.process.dao.PipelineSettingDao
+import com.tencent.devops.process.dao.label.PipelineGroupDao
+import com.tencent.devops.process.dao.label.PipelineLabelDao
+import com.tencent.devops.process.dao.label.PipelineViewDao
 import com.tencent.devops.process.engine.dao.PipelineInfoDao
 import com.tencent.devops.process.engine.service.PipelineRepositoryService
-import com.tencent.devops.process.service.PipelineInfoFacadeService
-import com.tencent.devops.store.api.template.ServiceTemplateResource
+import com.tencent.devops.process.pojo.classify.PipelineGroupCreate
+import com.tencent.devops.process.pojo.classify.PipelineLabelCreate
+import com.tencent.devops.process.pojo.classify.PipelineViewForm
+import com.tencent.devops.process.pojo.classify.enums.Logic
+import com.tencent.devops.process.service.label.PipelineGroupService
+import com.tencent.devops.process.service.view.PipelineViewService
 import org.jooq.DSLContext
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
@@ -25,40 +32,15 @@ class PipelineCopyService @Autowired constructor(
     private val pipelineInfoDao: PipelineInfoDao,
     private val pipelineSettingDao: PipelineSettingDao,
     private val pipelineRepositoryService: PipelineRepositoryService,
-    private val pipelineInfoFacadeService: PipelineInfoFacadeService,
     private val pipelineSettingFacadeService: PipelineSettingFacadeService,
+    private val pipelineGroupDao: PipelineGroupDao,
+    private val pipelineLabelDao: PipelineLabelDao,
+    private val pipelineGroupService: PipelineGroupService,
+    private val pipelineViewDao: PipelineViewDao,
+    private val pipelineViewService: PipelineViewService,
     private val client: Client,
     private val clientTokenService: ClientTokenService
 ) {
-
-    fun copyAcrossProject(
-        userId: String,
-        sourceProjectId: String,
-        targetProjectId: String,
-        pipelineId: String?
-    ) {
-        if (!pipelineId.isNullOrBlank()) {
-            try {
-                pipelineInfoDao.getPipelineInfo(
-                    dslContext = dslContext,
-                    projectId = sourceProjectId,
-                    pipelineId = pipelineId,
-                    delete = false
-                )?.let { sourcePipelineInfo ->
-                    copySinglePipeline(
-                        userId = userId,
-                        sourceProjectId = sourceProjectId,
-                        targetProjectId = targetProjectId,
-                        sourcePipelineInfo = sourcePipelineInfo
-                    )
-                } ?: logger.warn("get source pipeline failed|$sourceProjectId|$pipelineId")
-            } catch (ignored: Exception) {
-                logger.warn("get source pipeline failed|$sourceProjectId|$pipelineId", ignored)
-            }
-            return
-        }
-        copyAllPipelinesByPage(userId, sourceProjectId, targetProjectId)
-    }
 
     fun fixInstanceSetting(
         userId: String,
@@ -156,145 +138,57 @@ class PipelineCopyService @Autowired constructor(
         }
     }
 
-    private fun copyAllPipelinesByPage(
+    fun copyLabelsAcrossProject(
         userId: String,
         sourceProjectId: String,
-        targetProjectId: String
+        targetProjectId: String,
+        labelId: String?
     ) {
-        val pageSize = 100
-        var offset = 0
-        while (true) {
-            val page = pipelineInfoDao.listPipelineInfoByProject(
+        if (!labelId.isNullOrBlank()) {
+            try {
+                pipelineLabelDao.getById(
+                    dslContext = dslContext,
+                    projectId = sourceProjectId,
+                    id = HashUtil.decodeIdToLong(labelId)
+                )?.let { sourceLabel ->
+                    copySingleLabel(
+                        userId = userId,
+                        sourceProjectId = sourceProjectId,
+                        targetProjectId = targetProjectId,
+                        sourceLabel = sourceLabel
+                    )
+                } ?: logger.warn("get source pipeline label failed|$sourceProjectId|$labelId")
+            } catch (ignored: Exception) {
+                logger.warn("get source pipeline label failed|$sourceProjectId|$labelId", ignored)
+            }
+            return
+        }
+        copyAllLabelsByGroup(userId, sourceProjectId, targetProjectId)
+    }
+
+    fun copyViewsAcrossProject(
+        userId: String,
+        sourceProjectId: String,
+        targetProjectId: String,
+        viewName: String?
+    ) {
+        if (!viewName.isNullOrBlank()) {
+            pipelineViewDao.fetchAnyByName(
                 dslContext = dslContext,
                 projectId = sourceProjectId,
-                limit = pageSize,
-                offset = offset,
-                deleteFlag = false
-            ) ?: break
-            if (page.isEmpty()) {
-                break
-            }
-            page.forEach { sourcePipelineInfo ->
-                copySinglePipeline(
+                name = viewName,
+                isProject = true
+            )?.let { sourceView ->
+                copySingleView(
                     userId = userId,
                     sourceProjectId = sourceProjectId,
                     targetProjectId = targetProjectId,
-                    sourcePipelineInfo = sourcePipelineInfo
+                    sourceView = sourceView
                 )
-            }
-            if (page.size < pageSize) {
-                break
-            }
-            offset += pageSize
+            } ?: logger.warn("get source pipeline view failed|$sourceProjectId|$viewName")
+            return
         }
-    }
-
-    private fun copySinglePipeline(
-        userId: String,
-        sourceProjectId: String,
-        targetProjectId: String,
-        sourcePipelineInfo: TPipelineInfoRecord
-    ) {
-        val copiedPipelineId = sourcePipelineInfo.pipelineId
-        try {
-            if (shouldSkipPipelineCopy(
-                    targetProjectId = targetProjectId,
-                    pipelineId = copiedPipelineId,
-                    pipelineName = sourcePipelineInfo.pipelineName,
-                    channelCode = resolveChannelCode(sourcePipelineInfo)
-                )
-            ) {
-                logger.warn(
-                    "pipeline already exists, skip copy|$sourceProjectId|$targetProjectId|$copiedPipelineId"
-                )
-                return
-            }
-            val sourceResource = pipelineRepositoryService.getPipelineResourceVersion(
-                projectId = sourceProjectId,
-                pipelineId = copiedPipelineId
-            ) ?: run {
-                logger.warn(
-                    "source pipeline release version not found|$sourceProjectId|$copiedPipelineId"
-                )
-                return
-            }
-            val sourceSetting = pipelineRepositoryService.getSettingByPipelineVersion(
-                projectId = sourceProjectId,
-                pipelineId = copiedPipelineId,
-                pipelineVersion = sourceResource.version
-            ) ?: pipelineRepositoryService.getSetting(
-                projectId = sourceProjectId,
-                pipelineId = copiedPipelineId
-            )
-            val targetSetting = sourceSetting?.let {
-                copyPipelineSetting(
-                    sourceSetting = it,
-                    targetProjectId = targetProjectId,
-                    pipelineId = copiedPipelineId,
-                    pipelineName = sourcePipelineInfo.pipelineName
-                )
-            }
-            val pipelineOauthUser = pipelineRepositoryService.getPipelineOauthUser(
-                projectId = sourceProjectId,
-                pipelineId = copiedPipelineId
-            )
-            // 校验插件是否在目标项目可见
-            client.get(ServiceTemplateResource::class).validateModelComponentVisibleDept(
-                userId = userId,
-                model = sourceResource.model,
-                projectCode = targetProjectId
-            )
-            pipelineInfoFacadeService.createPipeline(
-                userId = pipelineOauthUser ?: userId,
-                projectId = targetProjectId,
-                model = sourceResource.model,
-                channelCode = resolveChannelCode(sourcePipelineInfo),
-                setting = targetSetting,
-                fixPipelineId = copiedPipelineId,
-                instanceType = if (sourceResource.model.instanceFromTemplate == true) {
-                    PipelineInstanceTypeEnum.CONSTRAINT.type
-                } else {
-                    PipelineInstanceTypeEnum.FREEDOM.type
-                },
-                versionStatus = sourceResource.status,
-                branchName = sourceResource.versionName?.takeIf {
-                    sourceResource.status == VersionStatus.BRANCH
-                }
-            )
-            copyPipelineGroupMembersSafely(
-                sourceProjectId = sourceProjectId,
-                targetProjectId = targetProjectId,
-                pipelineId = copiedPipelineId
-            )
-        } catch (ignored: Exception) {
-            logger.warn(
-                "copy pipeline failed|$sourceProjectId|$targetProjectId|$copiedPipelineId",
-                ignored
-            )
-        }
-    }
-
-    private fun shouldSkipPipelineCopy(
-        targetProjectId: String,
-        pipelineId: String,
-        pipelineName: String,
-        channelCode: ChannelCode
-    ): Boolean {
-        if (pipelineInfoDao.getPipelineInfo(
-                dslContext = dslContext,
-                projectId = targetProjectId,
-                pipelineId = pipelineId,
-                delete = false
-            ) != null
-        ) {
-            return true
-        }
-        return pipelineInfoFacadeService.isPipelineExist(
-            projectId = targetProjectId,
-            pipelineId = pipelineId,
-            name = pipelineName,
-            channelCode = channelCode
-        )
+        copyAllViewsByPage(userId, sourceProjectId, targetProjectId)
     }
 
     private fun copyPipelineSetting(
@@ -310,27 +204,202 @@ class PipelineCopyService @Autowired constructor(
         )
     }
 
-    private fun resolveChannelCode(sourcePipelineInfo: TPipelineInfoRecord): ChannelCode {
-        return ChannelCode.getChannel(sourcePipelineInfo.channel) ?: ChannelCode.BS
+    private fun copyAllLabelsByGroup(
+        userId: String,
+        sourceProjectId: String,
+        targetProjectId: String
+    ) {
+        pipelineGroupDao.list(dslContext, sourceProjectId).forEach { sourceGroup ->
+            pipelineLabelDao.getByGroupIds(
+                dslContext = dslContext,
+                projectId = sourceProjectId,
+                groupId = setOf(sourceGroup.id)
+            ).forEach { sourceLabel ->
+                copySingleLabel(
+                    userId = userId,
+                    sourceProjectId = sourceProjectId,
+                    targetProjectId = targetProjectId,
+                    sourceLabel = sourceLabel
+                )
+            }
+        }
     }
 
-    private fun copyPipelineGroupMembersSafely(
+    private fun copySingleLabel(
+        userId: String,
         sourceProjectId: String,
         targetProjectId: String,
-        pipelineId: String
+        sourceLabel: TPipelineLabelRecord
+    ) {
+        try {
+            val sourceGroup = pipelineGroupDao.get(
+                dslContext = dslContext,
+                projectId = sourceProjectId,
+                groupId = sourceLabel.groupId
+            ) ?: run {
+                logger.warn(
+                    "get source pipeline label group failed|$sourceProjectId|${sourceLabel.groupId}"
+                )
+                return
+            }
+            val targetGroup = getOrCreateTargetGroup(
+                userId = userId,
+                targetProjectId = targetProjectId,
+                groupName = sourceGroup.name
+            )
+            if (pipelineLabelDao.getByName(
+                    dslContext = dslContext,
+                    projectId = targetProjectId,
+                    groupId = targetGroup.id,
+                    name = sourceLabel.name
+                ) != null
+            ) {
+                logger.warn(
+                    "pipeline label already exists, skip copy|$sourceProjectId|$targetProjectId|${sourceLabel.name}"
+                )
+                return
+            }
+            pipelineGroupService.addLabel(
+                userId = userId,
+                projectId = targetProjectId,
+                pipelineLabel = PipelineLabelCreate(
+                    groupId = HashUtil.encodeLongId(targetGroup.id),
+                    name = sourceLabel.name
+                )
+            )
+        } catch (ignored: Exception) {
+            logger.warn(
+                "copy pipeline label failed|$sourceProjectId|$targetProjectId|${sourceLabel.name}",
+                ignored
+            )
+        }
+    }
+
+    private fun getOrCreateTargetGroup(
+        userId: String,
+        targetProjectId: String,
+        groupName: String
+    ) = pipelineGroupDao.getByName(
+        dslContext = dslContext,
+        projectId = targetProjectId,
+        name = groupName
+    ) ?: run {
+        pipelineGroupService.addGroup(
+            userId = userId,
+            pipelineGroup = PipelineGroupCreate(
+                projectId = targetProjectId,
+                name = groupName
+            )
+        )
+        pipelineGroupDao.getByName(
+            dslContext = dslContext,
+            projectId = targetProjectId,
+            name = groupName
+        ) ?: throw IllegalStateException("create pipeline label group failed|$targetProjectId|$groupName")
+    }
+
+    private fun copyAllViewsByPage(
+        userId: String,
+        sourceProjectId: String,
+        targetProjectId: String
+    ) {
+        val pageSize = 100
+        var offset = 0
+        while (true) {
+            val page = pipelineViewDao.listByPage(
+                dslContext = dslContext,
+                projectId = sourceProjectId,
+                isProject = true,
+                viewName = null,
+                limit = pageSize,
+                offset = offset
+            )
+            if (page.isEmpty()) {
+                break
+            }
+            page.forEach { sourceView ->
+                copySingleView(
+                    userId = userId,
+                    sourceProjectId = sourceProjectId,
+                    targetProjectId = targetProjectId,
+                    sourceView = sourceView
+                )
+            }
+            if (page.size < pageSize) {
+                break
+            }
+            offset += pageSize
+        }
+    }
+
+    private fun copySingleView(
+        userId: String,
+        sourceProjectId: String,
+        targetProjectId: String,
+        sourceView: TPipelineViewRecord
+    ) {
+        try {
+            if (pipelineViewDao.fetchAnyByName(
+                    dslContext = dslContext,
+                    projectId = targetProjectId,
+                    name = sourceView.name,
+                    isProject = true
+                ) != null
+            ) {
+                logger.warn(
+                    "pipeline view already exists, skip copy|$sourceProjectId|$targetProjectId|${sourceView.name}"
+                )
+                return
+            }
+            val targetViewId = pipelineViewService.addView(
+                userId = userId,
+                projectId = targetProjectId,
+                pipelineView = PipelineViewForm(
+                    name = sourceView.name,
+                    projected = true,
+                    viewType = sourceView.viewType,
+                    logic = Logic.of(sourceView.logic),
+                    filters = pipelineViewService.getFilters(
+                        filterByName = sourceView.filterByPipeineName,
+                        filterByCreator = sourceView.filterByCreator,
+                        filters = sourceView.filters
+                    ),
+                    pipelineIds = if (sourceView.viewType == PipelineViewType.STATIC) emptyList() else null
+                )
+            )
+            copyViewGroupMembersSafely(
+                sourceProjectId = sourceProjectId,
+                targetProjectId = targetProjectId,
+                sourceViewId = sourceView.id,
+                targetViewId = targetViewId
+            )
+        } catch (ignored: Exception) {
+            logger.warn(
+                "copy pipeline view failed|$sourceProjectId|$targetProjectId|${sourceView.name}",
+                ignored
+            )
+        }
+    }
+
+    private fun copyViewGroupMembersSafely(
+        sourceProjectId: String,
+        targetProjectId: String,
+        sourceViewId: Long,
+        targetViewId: Long
     ) {
         try {
             client.get(ServiceResourceMemberResource::class).copyResourceGroupMembers(
                 token = clientTokenService.getSystemToken() ?: "",
                 sourceProjectCode = sourceProjectId,
-                resourceType = AuthResourceType.PIPELINE_DEFAULT.value,
-                sourceResourceCode = pipelineId,
-                targetProjectCode = targetProjectId,
-                targetResourceCode = pipelineId
+                resourceType = AuthResourceType.PIPELINE_GROUP.value,
+                sourceResourceCode = HashUtil.encodeLongId(sourceViewId),
+                targetResourceCode = HashUtil.encodeLongId(targetViewId),
+                targetProjectCode = targetProjectId
             )
         } catch (ignored: Exception) {
             logger.warn(
-                "copy pipeline group members failed|$sourceProjectId|$targetProjectId|$pipelineId",
+                "copy pipeline group members failed|$sourceProjectId|$targetProjectId|" +
+                    "$sourceViewId|$targetViewId",
                 ignored
             )
         }
